+++ application/controllers/Home.php
@@ -6,27 +6,16 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 	Home controller, uses data from many models (flights, and fleets)
 */
 class Home extends Application {
-
-    private static $SESS_PATH =  APPPATH . 'tmp/';
-
-    public function __construct () {
-        parent::__construct ();
-        var_dump (Home::$SESS_PATH);
-        ini_set('session.save_path', Home::$SESS_PATH);
-        $this->load->library ('session');
-        var_dump ($this -> session);
-        $arr = array ('123');
-        $this->session -> item = $arr;
-        var_dump($this -> session -> item);
-        session_destroy();
-    }
-
+	
 	/**
 	 * Index Page for this controller.
 	 */
 	public function index()
 	{
 		$this->data['pagebody']   = 'welcome_message';
+		$role = $this->session->userdata('userrole');
+		$this->data['pagetitle'] = ' Welcome to Albatros Airlines! ('. $role . ')';
+
 		$airline = $this -> flights -> get_flight_airline () -> getViewArray ();
 		
         $this -> data['airports'] = array (