+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2017-11-12 - YR
+### Added 
+- Added bootstrap css and js files
+- User roles to application
+
+### Updated
+- Template view php to use bootstrap
+- Added pagetitle to template view
+
 ## 2017-11-12 - NS
 ### Added 
 - Sessions and a temp